Adding Frontend updateProduct

+++ backend/src/main/java/org/example/backend/dto/ProductDTO.java
@@ -2,6 +2,6 @@ package org.example.backend.dto;
 
 public record ProductDTO(
         String name,
-        int price
+        float price
 ) {
 }
