+++ backend/src/main/java/org/example/backend/model/Product.java
@@ -8,7 +8,7 @@ public record Product(
         @Id
         String id,
         String name,
-        int price
+        float price
 
 ) {
 }
